Improved Ball VFX
I had to revamp the entire collision system a bit, just so placed trinkets could be easily identifiable and trinkets that were already on the map would fit into that. Also added a few more materials and added in two dynamic VFX systems into the balls. Type 1 is active based on ball speed and type 2 is active based on the number of collisions in a short timespan. This obviously resets and has a bit of leniency so the fire doesn't immediately disappear.